requirements.txt
aiogram==2.8
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.5.1
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.18.2
pandas==1.0.3
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
yarl==1.4.2
